requirements.txt
python-dotenv==1.0.1
python-telegram-bot[callback-data]==20.7
